Fix 32-bit issues
On arm32-centos7, the build fails
in stub_libstore.cc and in src/security/PeerOptions.cc
because size_t and long are both 32 bit

+++ src/tests/stub_libstore.cc
@@ -145,7 +145,7 @@ StoreSearch *NewLocalSearch() STUB_RETVAL(nullptr)
 
 #include "store/SwapMetaIn.h"
 size_t Store::UnpackSwapMetaSize(const SBuf &) STUB_RETVAL(0)
-uint64_t Store::UnpackIndexSwapMeta(const MemBuf &, StoreEntry &, cache_key *) STUB_RETVAL(0)
+size_t Store::UnpackIndexSwapMeta(const MemBuf &, StoreEntry &, cache_key *) STUB_RETVAL(0)
 void Store::UnpackHitSwapMeta(char const *, ssize_t, StoreEntry &) STUB
 
 #include "store/SwapMetaOut.h"
